+++ frontend/package.json
@@ -1,28 +1,28 @@
 {
-  "name": "muse-ant-design-dashboard",
-  "version": "1.0.0",
+  "name": "frontend",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^6.0.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "antd": "^5.4.2",
-    "apexcharts": "^4.5.0",
-    "axios": "^1.8.4",
-    "cross-env": "^7.0.3",
-    "install": "^0.13.0",
-    "react": "19.1.0",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.126",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
+    "antd": "^5.24.6",
+    "react": "^19.1.0",
     "react-apexcharts": "^1.7.0",
-    "react-dom": "19.1.0",
+    "react-dom": "^19.1.0",
     "react-router-dom": "^7.5.0",
-    "react-scripts": "^5.0.1",
-    "styled-components": "^5.3.0",
-    "web-vitals": "^1.1.2"
+    "react-scripts": "5.0.1",
+    "typescript": "^5.8.3",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "cross-env CI=false cross-env NODE_OPTIONS=--openssl-legacy-provider react-scripts build",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -43,9 +43,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@babel/core": "^7.16.0"
   },
   "proxy": "http://localhost:8090"
 }
