+++ frontend/public/manifest.json
@@ -1,14 +1,14 @@
 {
-  "short_name": "Muse Ant Design Dashboard",
-  "name": "Muse Ant Design Dashboard",
+  "short_name": "React App",
+  "name": "Create React App Sample",
   "icons": [
     {
-      "src": "favicon.png",
+      "src": "favicon.ico",
       "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/png"
+      "type": "image/x-icon"
     },
     {
-      "src": "favicon.png",
+      "src": "logo192.png",
       "type": "image/png",
       "sizes": "192x192"
     },
