+++ frontend/src/components/layout/Header.js
@@ -36,7 +36,7 @@ import {
 
 import { NavLink, Link } from "react-router-dom";
 import styled from "styled-components";
-import avtar from "../../assets/images/team-2.jpg";
+import avtar from "../../../../../typetest/frontend/src/assets/images/team-2.jpg";
 
 const ButtonContainer = styled.div`
   .ant-btn-primary {
