--- frontend/src/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-import * as ReactDOM from 'react-dom/client';
-import App from "./App";
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-    <>
-        <App />
-    </>
-);
