+++ frontend/src/pages/Detail.tsx
@@ -1,16 +1,13 @@
 import {useParams} from "react-router-dom";
 
 function Detail() {
-
     const {id} = useParams();
-
     return (
         <>
-           <span>tttt</span>
-           <span>{id}</span>
+            <span>tttt</span>
+            <span>{id}</span>
         </>
     )
 }
-
 
 export default Detail;
